Break out views.py into smaller files
