+++ companies/views/home.py
@@ -1,0 +1,4 @@
+from django.shortcuts import render
+
+def home(request):
+    return render(request, 'companies/home.html')
